config.py
import os
import sys
from dotenv import load_dotenv

basedir = os.path.abspath(os.path.dirname(__file__))
load_dotenv(os.path.join(basedir,'.env'))

WIN = sys.platform.startswith('win')
if WIN: 
    prefix = 'sqlite:///'
else:  
    prefix = 'sqlite:////'
    
class Config(object):
    SQLALCHEMY_DATABASE_URI =  os.environ.get('DATABASE_URL') or 'sqlite:///'+os.path.join(basedir,'app.db')
    SQLALCHEMY_TRACK_MODIFICATIONS = False