requirements.txt
alembic==1.8.1
click==8.1.3
dominate==2.7.0
Flask==2.2.2
Flask-Admin==1.6.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.2
MarkupSafe==2.1.1
python-dotenv==0.20.0
SQLAlchemy==1.4.40
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
